requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.1
chromadb>=0.5.0
sentence-transformers>=2.5.1
numpy>=1.24
python-multipart>=0.0.9
openai-whisper>=20231117
opencv-python>=4.9.0
mediapipe>=0.10.0
pillow>=10.0.0
transformers>=4.39.0
celery>=5.3.0
redis>=5.0.0
llama-cpp-python>=0.2.60
